+++ src/assets/img/index.ts
@@ -1,2 +1,3 @@
 export { default as border } from "./BORDER.png";
 export { default as book } from "./book.png";
+export { default as line } from "./Line.png";
