--- src/pages/Home/BookInfoPage.css
@@ -1,357 +1,0 @@
-.bookinfo-frame {
-    display: flex;
-    flex-direction: column;
-    width: 1040px;
-    height: 2815px;
-    align-items: flex-start;
-    justify-content: space-between;
-    padding: 80px 0px 100px;
-    position: relative;
-}
-
-.bookinfo-frame .bookinfo-div {
-    display: flex;
-    flex-direction: column;
-    width: 1040px;
-    align-items: flex-start;
-    gap: 72px;
-    position: relative;
-    flex: 0 0 auto;
-}
-
-.bookinfo-frame .bookinfo-group {
-    position:relative;
-    width: 588px;
-    height: 37px;
-}
-
-.bookinfo-frame .bookinfotext-wrapper {
-    position: absolute;
-    width: 586px;
-    height: 37px;
-    top: 0;
-    left: 0;
-    font-family: "Pretendard-SemiBold", Helvetica;
-    font-weight: 700;
-    color: #7b9acc;
-    font-size: 40px;
-    letter-spacing: 0;
-    line-height: normal;
-    white-space: nowrap;
-}
-
-.bookinfo-frame .bookinfo-div2 {
-    display: flex;
-    flex-wrap: wrap;
-    align-items: flex-start;
-    gap: 70px 82px;
-    position: relative;
-    align-self: stretch;
-    width: 100%;
-    flex: 0 0 auto;
-}
-
-/* 책 테두리 */
-.bookinfo-frame .bookinfo-view {
-    display: flex;
-    width: 310px;
-    height: 400px;
-    align-items: center;
-    justify-content: center;
-    gap: 10px;
-    padding: 50px 51px;
-    position: relative;
-    background-color: #ffffff;
-    border-radius: 20px;
-    border: 2px solid;
-    border-color: #d4d4d4;
-}
-
-/* 책 이미지 */
-.bookinfo-frame .bookinfo-image {
-    position: relative;
-    width:214.75px;
-    height: 308px;
-    margin-bottom: -8.00px;
-    margin-left: -3.38px;
-    margin-right: -3.38px;
-    object-fit: cover;
-}
-
-.bookinfo-frame .bookinfo-div3 {
-    display: flex;
-    flex-direction: column;
-    width: 498px;
-    align-items: flex-start;
-    gap: 29px;
-    position: relative;
-}
-
-.bookinfo-frame .bookinfo-div4 {
-    display: flex;
-    flex-direction: column;
-    width: 143px;
-    align-items: flex-start;
-    gap: 9px;
-    position: relative;
-    flex: 0 0 auto;
-}
-
-/* 책 제목 */
-.bookinfo-frame .bookinfotext-wrapper2 {
-    position: relative;
-    align-self: stretch;
-    height: 56px;
-    margin-top:  -1.00px;
-    font-family: "Pretendard-SemiBold", Helvetica;
-    font-weight: 600;
-    color: #0e2624;
-    font-size: 30px;
-    letter-spacing: 0;
-    line-height: 29.9px;
-}
-
-/* 작가 */
-.bookinfo-frame .bookinfotext-wrapper3 {
-    position: relative;
-    align-self: stretch;
-    height: 28px;
-    font-family: "Pretendard-SemiBold", Helvetica;
-    font-weight: 700;
-    color: #3f4756;
-    font-size: 24px;
-    letter-spacing: 0;
-    line-height: 29.9px;
-    white-space: nowrap;
-}
-
-/* 출판사 및 출판 날짜 */
-.bookinfo-frame .bookinfo-p {
-    position: relative;
-    align-self: stretch;
-    font-family: "Pretendard-SemiBold", Helvetica;
-    font-weight: 400;
-    color: #666666;
-    font-size: 24px;
-    letter-spacing: 0;
-    line-height: normal;
-}
-
-.bookinfo-frame .bookinfo-group-wrapper {
-    position: relative;
-    width: 1040px;
-    height: 214.36px;
-}
-
-/* 채팅방 참여하기, 채팅방 규칙 */
-.bookinfo-frame .bookinfo-chat-room-join-wrapper {
-    display: flex;
-    flex-direction: column;
-    width: 1042px;
-    height: 216px;
-    align-items: center;
-    justify-content: center;
-    gap: 10px;
-    padding: 25px 40px;
-    position: relative;
-    top: -1px;
-    left: -1px;
-    border-radius: 5px;
-    border: 1px solid;
-    border-color: #d4d4d4;
-}
-
-.bookinfo-frame .bookinfo-div5 {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 110px;
-    position: relative;
-    align-self: stretch;
-    width: 100%;
-    flex: 0 0 auto;
-}
-
-/* {책 제목} 채팅방 참여하기, 여러사람들과 {책제목}에 대해 이야기 해봅시다. */
-.bookinfo-frame .bookinfo-div6 {
-    display: flex;
-    flex-direction: column;
-    width: 393px;
-    align-items: flex-start;
-    gap: 14px;
-    position: relative;
-}
-
-/* {책 제목} 채팅방 참여하기 */
-.bookinfo-frame .bookinfotext-wrapper4 {
-    position: relative;
-    width: 287px;
-    height: 29px;
-    margin-top: -1.00px;
-    font-family: "Pretendard-ExtraBold", Helvetica;
-    font-weight: 800;
-    color: #3f4756;
-    font-size: 25px;
-    white-space:  nowrap;
-    letter-spacing: 0;
-    line-height: normal;
-}
-
-/* 여러 사람들과 {책 제목}에 대해 이야기 해봅시다. */
-.bookinfo-frame .bookinfotext-wrapper5 {
-    position: relative;
-    align-self: stretch;
-    height: 19px;
-    font-family: "Pretendard-Regular", Helvetica;
-    font-weight: 400;
-    color: #3f4756;
-    font-size: 16px;
-    white-space: nowrap;
-    letter-spacing: 0;
-    line-height: normal;
-}
-
-/* 토론방 규칙, 규칙 설명, 참여하기 버튼 */
-.bookinfo-frame .bookinfo-div7 {
-    display: flex;
-    flex-direction: column;
-    width: 400px;
-    align-items: flex-start;
-    gap: 21px;
-    position: relative;
-}
-
-.bookinfo-frame .bookinfo-div8 {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    position: relative;
-    align-self: stretch;
-    width: 100%;
-    flex: 0 0 auto;
-}
-
-/* 토론방 규칙 */
-.bookinfo-frame .bookinfotext-wrapper6 {
-    position:relative;
-    align-self: stretch;
-    height: 87.36px;
-    margin-top:  -1.00px;
-    font-family: "Pretendard-Regular", Helvetica;
-    font-weight: 400;
-    color: #3f4756;
-    font-size: 16px;
-    letter-spacing: 0;
-    line-height: normal;
-}
-
-/* 규칙 상세 설명 */
-.bookinfo-frame .bookinfotext-wrapper7 {
-    position: relative;
-    align-self: stretch;
-    height: 51px;
-    margin-top: -60px;
-    font-family: "Pretendard-Light", Helvetica;
-    font-weight: 300;
-    color: #3f4756;
-    font-size: 14px;
-    letter-spacing: 0;
-    line-height: normal;
-}
-
-/* 채팅방 참여하기 버튼 테두리 */
-.bookinfo-frame .bookinfo-div-wrapper {
-    display: flex;
-    width: 345px;
-    height: 56px;
-    align-items: center;
-    justify-content: center;
-    gap: 10px;
-    padding: 11px 32px;
-    position: relative;
-    background-color: #7b9acc;
-    border-radius: 10px;
-}
-
-/* 책 소개, 투표, 출판사 서평  큰 틀*/
-.bookinfo-frame .bookinfo-div9 {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 72px;
-    position: relative;
-    align-self: stretch;
-    width: 100%;
-    flex: 0 0 auto;
-}
-
-.bookinfo-frame .bookinfo-div10 {
-    display: flex;
-    flex-direction: column;
-    width: 990.34px;
-    align-items: flex-start;
-    gap: 11px;
-    position: relative;
-    flex: 0 0 auto;
-}
-
-/* 책 소개 투표 출판사 서평 틀 */
-.bookinfo-frame .bookinfo-div11 {
-    display: inline-flex;
-    align-items: center;
-    gap: 50px;
-    position: relative;
-    flex: 0 0 auto;
-}
-
-/* 클릭한 부분의 폰트색상 #3f4756으로 바뀌게 만들어야 함 */
-.bookinfo-frame .bookinfotext-wrapper9 {
-    position: relative;
-    width: 76px;
-    height: 29px;
-    margin-top: -1.00px;
-    font-family: "Pretendard-Medium", Helvetica;
-    font-weight: 500;
-    color: #3f4756;
-    font-size: 22px;
-    letter-spacing: 0;
-    line-height: normal;
-}
-
-/* 이게 기본 상태 */
-.bookinfo-frame .bookinfotext-wrapper10 {
-    position: relative;
-    width: 49px;
-    height: 29px;
-    margin-top: -1.00px;
-    font-family: "Pretendard-Medium", Helvetica;
-    font-weight: 500;
-    color: #666666;
-    font-size: 22px;
-    letter-spacing: 0;
-    line-height: normal;
-}
-
-.bookinfo-frame .bookinfotext-wrapper11 {
-    position: relative;
-    width: 117px;
-    height: 29px;
-    margin-top: -1.00px;
-    font-family: "Pretendard-Medium", Helvetica;
-    font-weight: 500;
-    color: #666666;
-    font-size: 22px;
-    letter-spacing: 0;
-    line-height: normal;
-}
-/* bar line 일단은 image로 넣어놨습니다. */
-.bookinfo-frame .bookinfo-line {
-    position: relative;
-    width: 990.34px;
-    height: 2.5px;
-    margin-bottom: -1.00px;
-}
-
-.bookinfo-frame .bookinfo-description-wrapper
-
-
